Working on the buying api, changed wallet in models

+++ models.py
@@ -11,7 +11,7 @@ class User(db.Model):
     password = db.Column(db.String(100))
     admin = db.Column(db.BOOLEAN, default=False)
     online = db.Column(db.BOOLEAN, default=False)
-    current_balance = db.Column(db.Integer, default=1000)
+    current_balance = db.Column(db.Float, default=1000)
 
     def is_active(self):
         return True
@@ -41,9 +41,9 @@ class TransactionHistory(db.Model):
     __tablename__ = "transaction_history"
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     wallet_id = db.Column(db.Integer, db.ForeignKey('wallet.id'), nullable=False)
-    amount_usd = db.Column(db.Integer)
+    amount_usd = db.Column(db.Float)
     token_name = db.Column(db.String(150))
-    token_amount = db.Column(db.Integer)
+    token_amount = db.Column(db.Float)
     time = db.Column(db.DateTime, default=datetime.datetime.now())
